backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Task queue
celery==5.3.4
redis==5.0.1

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
chardet==5.2.0

# AI/LLM
anthropic==0.7.8

# Environment and config
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Utilities
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1